Implement rollback, persist order details.

+++ app/Http/Controllers/OrderController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Bill;
 use App\Customer;
 use App\Order;
+use App\OrderDetail;
 use DateTime;
 use Illuminate\Http\Request;
 
@@ -44,11 +45,11 @@ class OrderController extends Controller
         if($bill->isValid($billRequest)){
             $saved = $bill->persist($billRequest);
             if(!$saved) {
-                $customer->delete();
+                $this->rollback($customer);
                 return response()->json([ "errorMessage" => "Error on save bill"], 500);
             }
         } else {
-            $customer->delete();
+            $this->rollback($customer);
             return response()->json([ "errorMessage" => "Bad Request"], 400);
         }
 
@@ -58,15 +59,38 @@ class OrderController extends Controller
         $order->date = date('Y-m-d H:i:s');
         $saved = $order->save();
         if(!$saved) {
-            $customer->delete();
-            $bill->delete();
+            $this->rollback($customer, $bill);
             return response()->json([ "errorMessage" => "Error on save order"], 500);
         }
 
         $productsRequest = json_decode(json_encode($request->products));
+        $orderDetailList = [];
 
+        foreach ($productsRequest as $product) {
+            $orderDetail = new OrderDetail();
+            $orderDetail->id_order = $order->id;
+            if($orderDetail->isValid($product)){
+                $saved = $orderDetail->persist($product);
+                if(!$saved) {
+                    $this->rollback($customer, $bill, $orderDetailList);
+                    return response()->json([ "errorMessage" => "Error on save order details"], 500);
+                }
+                array_push($orderDetailList, $orderDetail);
+            } else {
+                $this->rollback($customer, $bill, $orderDetailList);
+                return response()->json([ "errorMessage" => "Bad Request"], 400);
+            }
+        }
 
         return response()->json([ "message" => "Success"], 201);
+    }
+
+    private function rollback($customer, $bill = null, $orderDetailList = []) {
+        $customer && $customer->delete();
+        $bill && $bill->delete();
+        foreach ($orderDetailList as $orderDetail) {
+            $orderDetail->delete();
+        }
     }
 
     /**
